requirements.txt
streamlit==1.32.0
pandas==2.0.3
matplotlib==3.7.1
openai==1.30.1
google-generativeai==0.3.2
sweetviz==2.1.4
openpyxl==3.1.2